rocklist: prepend an absolute partial path instead of a relative one
For some reason, if a playlist was used by shuffling it and inserting it into
the 'current playlist', only the first song would play, and the rest would show
'(ERR)' before them in the playlist screen (i.e. not found).

Using an absolute path is simple for my setup and I'd only need to change it if
I move the location of the music folder on my Rockbox device.